Move SentimentDatum to classifiers/data.py and make the reading code read in SentimentDatum
