+++ stanza/models/classifiers/data.py
@@ -1,6 +1,7 @@
 """Stanza models classifier data functions."""
 
 import collections
+from collections import namedtuple
 import logging
 import json
 import random
@@ -12,6 +13,7 @@ from stanza.models.common.vocab import PAD, PAD_ID, UNK, UNK_ID
 
 logger = logging.getLogger('stanza')
 
+SentimentDatum = namedtuple('SentimentDatum', ['sentiment', 'text'])
 
 def update_text(sentence: List[str], wordvec_type: WVType) -> List[str]:
     """
@@ -46,7 +48,7 @@ def update_text(sentence: List[str], wordvec_type: WVType) -> List[str]:
         raise ValueError("Unknown wordvec_type {}".format(wordvec_type))
 
 
-def read_dataset(dataset, wordvec_type: WVType, min_len: int) -> List[tuple]:
+def read_dataset(dataset, wordvec_type: WVType, min_len: int) -> List[SentimentDatum]:
     """
     returns a list where the values of the list are
       label, [token...]
@@ -60,16 +62,16 @@ def read_dataset(dataset, wordvec_type: WVType, min_len: int) -> List[tuple]:
     # TODO: maybe do this processing later, once the model is built.
     # then move the processing into the model so we can use
     # overloading to potentially make future model types
-    lines = [(x[0], update_text(x[1], wordvec_type)) for x in lines]
+    lines = [SentimentDatum(x[0], update_text(x[1], wordvec_type)) for x in lines]
     if min_len:
-        lines = [x for x in lines if len(x[1]) >= min_len]
+        lines = [x for x in lines if len(x.text) >= min_len]
     return lines
 
 def dataset_labels(dataset):
     """
     Returns a sorted list of label name
     """
-    labels = set([x[0] for x in dataset])
+    labels = set([x.sentiment for x in dataset])
     if all(re.match("^[0-9]+$", label) for label in labels):
         # if all of the labels are integers, sort numerically
         # maybe not super important, but it would be nicer than having
@@ -82,7 +84,7 @@ def dataset_labels(dataset):
 def dataset_vocab(dataset):
     vocab = set()
     for line in dataset:
-        for word in line[1]:
+        for word in line.text:
             vocab.add(word)
     vocab = [PAD, UNK] + list(vocab)
     if vocab[PAD_ID] != PAD or vocab[UNK_ID] != UNK:
@@ -96,11 +98,11 @@ def sort_dataset_by_len(dataset):
     an OrderedDict is used so that the mapping is sorted from smallest to largest
     """
     sorted_dataset = collections.OrderedDict()
-    lengths = sorted(list(set(len(x[1]) for x in dataset)))
+    lengths = sorted(list(set(len(x.text) for x in dataset)))
     for l in lengths:
         sorted_dataset[l] = []
     for item in dataset:
-        sorted_dataset[len(item[1])].append(item)
+        sorted_dataset[len(item.text)].append(item)
     return sorted_dataset
 
 def shuffle_dataset(sorted_dataset):
