+++ stanza/utils/datasets/sentiment/process_MELD.py
@@ -28,7 +28,7 @@ import csv
 import os
 import sys
 
-from stanza.utils.datasets.sentiment.process_utils import SentimentDatum
+from stanza.models.classifiers.data import SentimentDatum
 import stanza.utils.datasets.sentiment.process_utils as process_utils
 
 def get_phrases(in_filename):
