+++ stanza/utils/datasets/sentiment/process_ren_chinese.py
@@ -9,7 +9,7 @@ from collections import namedtuple
 
 import stanza
 
-from stanza.utils.datasets.sentiment.process_utils import SentimentDatum
+from stanza.models.classifiers.data import SentimentDatum
 import stanza.utils.datasets.sentiment.process_utils as process_utils
 
 """
