+++ stanza/utils/datasets/sentiment/process_slsd.py
@@ -22,7 +22,7 @@ process_slsd.py <directory> <outputfile>
 import os
 import sys
 
-from stanza.utils.datasets.sentiment.process_utils import SentimentDatum
+from stanza.models.classifiers.data import SentimentDatum
 import stanza.utils.datasets.sentiment.process_utils as process_utils
 
 def get_phrases(in_directory):
