+++ stanza/utils/datasets/sentiment/process_sst.py
@@ -2,7 +2,7 @@ import argparse
 import os
 import subprocess
 
-from stanza.utils.datasets.sentiment.process_utils import SentimentDatum
+from stanza.models.classifiers.data import SentimentDatum
 import stanza.utils.datasets.sentiment.process_utils as process_utils
 
 import stanza.utils.default_paths as default_paths
